TIP-221: move filter dumpers to the command
