+++ src/Pim/Bundle/CatalogBundle/Command/DumperInterface.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Pim\Bundle\CatalogBundle\Query\Filter;
+namespace Pim\Bundle\CatalogBundle\Command;
 
 use Symfony\Component\Console\Helper\HelperSet;
 use Symfony\Component\Console\Output\OutputInterface;
